🎨  Move `scheduling` and `storage` in `adapters/`
no issue

Move `core/server/scheduling` to `core/server/adapters/scheduling` and
`core/server/storage` to `core/server/adapters/storage`
