+++ core/server/adapters/scheduling/SchedulingDefault.js
@@ -2,8 +2,8 @@ var util = require('util'),
     moment = require('moment'),
     request = require('superagent'),
     SchedulingBase = require(__dirname + '/SchedulingBase'),
-    errors = require(__dirname + '/../errors'),
-    logging = require(__dirname + '/../logging');
+    errors = require(__dirname + '/../../errors'),
+    logging = require(__dirname + '/../../logging');
 
 /**
  * allJobs is a sorted list by time attribute
