+++ core/server/adapters/scheduling/post-scheduling/index.js
@@ -1,10 +1,10 @@
 var Promise = require('bluebird'),
     moment = require('moment'),
     localUtils = require(__dirname + '/../utils'),
-    events = require(__dirname + '/../../events'),
-    errors = require(__dirname + '/../../errors'),
-    models = require(__dirname + '/../../models'),
-    schedules = require(__dirname + '/../../api/schedules'),
+    events = require(__dirname + '/../../../events'),
+    errors = require(__dirname + '/../../../errors'),
+    models = require(__dirname + '/../../../models'),
+    schedules = require(__dirname + '/../../../api/schedules'),
     _private = {};
 
 _private.normalize = function normalize(options) {
