+++ core/server/index.js
@@ -27,7 +27,7 @@ var debug = require('debug')('ghost:boot:init'),
     xmlrpc = require('./data/xml/xmlrpc'),
     slack = require('./data/slack'),
     GhostServer = require('./ghost-server'),
-    scheduling = require('./scheduling'),
+    scheduling = require('./adapters/scheduling'),
     settings = require('./settings'),
     settingsCache = require('./settings/cache'),
     themes = require('./themes'),
