+++ core/test/unit/adapters/scheduling/index_spec.js
@@ -2,8 +2,8 @@ var should = require('should'), // jshint ignore:line
     sinon = require('sinon'),
     rewire = require('rewire'),
     Promise = require('bluebird'),
-    config = require(__dirname + '/../../../server/config'),
-    postScheduling = require(__dirname + '/../../../server/scheduling/post-scheduling'),
+    config = require(__dirname + '/../../../../server/config'),
+    postScheduling = require(__dirname + '/../../../../server/adapters/scheduling/post-scheduling'),
 
     sandbox = sinon.sandbox.create();
 
@@ -12,7 +12,7 @@ describe('Scheduling', function () {
 
     before(function () {
         sandbox.stub(postScheduling, 'init').returns(Promise.resolve());
-        scope.scheduling = rewire(config.get('paths').corePath + '/server/scheduling');
+        scope.scheduling = rewire(config.get('paths').corePath + '/server/adapters/scheduling');
     });
 
     after(function () {
