+++ core/test/unit/adapters/scheduling/post-scheduling/index_spec.js
@@ -2,15 +2,15 @@ var should = require('should'),
     sinon = require('sinon'),
     Promise = require('bluebird'),
     moment = require('moment'),
-    config = require(__dirname + '/../../../../server/config'),
+    config = require(__dirname + '/../../../../../server/config'),
     testUtils = require(config.get('paths').corePath + '/test/utils'),
     errors = require(config.get('paths').corePath + '/server/errors'),
     events = require(config.get('paths').corePath + '/server/events'),
     models = require(config.get('paths').corePath + '/server/models'),
     api = require(config.get('paths').corePath + '/server/api'),
-    schedulingUtils = require(config.get('paths').corePath + '/server/scheduling/utils'),
-    SchedulingDefault = require(config.get('paths').corePath + '/server/scheduling/SchedulingDefault'),
-    postScheduling = require(config.get('paths').corePath + '/server/scheduling/post-scheduling'),
+    schedulingUtils = require(config.get('paths').corePath + '/server/adapters/scheduling/utils'),
+    SchedulingDefault = require(config.get('paths').corePath + '/server/adapters/scheduling/SchedulingDefault'),
+    postScheduling = require(config.get('paths').corePath + '/server/adapters/scheduling/post-scheduling'),
 
     sandbox = sinon.sandbox.create();
