+++ core/test/unit/adapters/scheduling/utils_spec.js
@@ -1,14 +1,14 @@
 var should = require('should'),
     fs = require('fs'),
-    config = require(__dirname + '/../../../server/config'),
+    config = require(__dirname + '/../../../../server/config'),
     errors = require(config.get('paths').corePath + '/server/errors'),
-    schedulingUtils = require(config.get('paths').corePath + '/server/scheduling/utils');
+    schedulingUtils = require(config.get('paths').corePath + '/server/adapters/scheduling/utils');
 
 describe('Scheduling: utils', function () {
     describe('success', function () {
         it('create good adapter', function (done) {
             schedulingUtils.createAdapter({
-                active: __dirname + '/../../../server/scheduling/SchedulingDefault'
+                active: __dirname + '/../../../../server/adapters/scheduling/SchedulingDefault'
             }).then(function (adapter) {
                 should.exist(adapter);
                 done();
@@ -18,7 +18,7 @@ describe('Scheduling: utils', function () {
         it('create good adapter', function (done) {
             var jsFile = '' +
                 'var util = require(\'util\');' +
-                'var SchedulingBase = require(__dirname + \'/../../../server/scheduling/SchedulingBase\');' +
+                'var SchedulingBase = require(__dirname + \'/../../../../server/adapters/scheduling/SchedulingBase\');' +
                 'var AnotherAdapter = function (){ SchedulingBase.call(this); };' +
                 'util.inherits(AnotherAdapter, SchedulingBase);' +
                 'AnotherAdapter.prototype.run = function (){};' +
@@ -62,7 +62,7 @@ describe('Scheduling: utils', function () {
         it('create with adapter, but missing fn\'s', function (done) {
             var jsFile = '' +
                 'var util = require(\'util\');' +
-                'var SchedulingBase = require(__dirname + \'/../../../server/scheduling/SchedulingBase\');' +
+                'var SchedulingBase = require(__dirname + \'/../../../../server/adapters/scheduling/SchedulingBase\');' +
                 'var BadAdapter = function (){ SchedulingBase.call(this); };' +
                 'util.inherits(BadAdapter, SchedulingBase);' +
                 'BadAdapter.prototype.schedule = function (){};' +
